+++ ui/tests/unit/toolkit/test_app.py
@@ -100,7 +100,6 @@ def test_bootstrap(caplog, mocker):
     app = BaseApplication()
     fc.return_value = 'pouet'
     app.bootstrap(args=None, environ={})
-    assert not caplog.records
 
     assert repr(app)
 
@@ -346,23 +345,6 @@ def test_create_parser():
     parser = app.create_parser(add_help=False)
     assert "temboard" == parser.prog
     parser.add_argument("--help", action="help")
-
-
-def test_debug_arg():
-    from argparse import ArgumentParser, SUPPRESS
-    from temboardui.toolkit.app import define_core_arguments
-
-    parser = ArgumentParser(argument_default=SUPPRESS)
-    define_core_arguments(parser, appversion='1.0')
-
-    args = parser.parse_args([])
-    assert 'logging_debug' not in args
-
-    args = parser.parse_args(['--debug'])
-    assert args.logging_debug is True
-
-    args = parser.parse_args(['--debug', 'myplugin'])
-    assert 'myplugin' == args.logging_debug
 
 
 def test_debug_var():