+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This release requires a database schema migration. Use temboard-migratedb
 script to proceed.
 
 - New database migration engine.
+- Removed parameter of `temboard --debug` CLI option.
+- Move `temboard-migratedb` as `temboard migratedb` subcommand.
+- New subcommand `temboard runtask` for debugging.
 
 
 ## [7.11] - Unreleased